removed adapter tests
